feat: add demo macro and split widget demos
